refactoring:delete ChatMessage
