+++ telegram_framework/messages/__init__.py
@@ -1,5 +1,4 @@
 from .text_message import Message, create_message
 from .reply import Reply, create_reply
-from .chat_message import ChatMessage, create_chat_message
-from .chat_message_base import is_chat_message
+from .chat_message import create_chat_message, is_chat_message
 from .media import Image, create_image
