+++ telegram_framework/messages/core/core_message_default.py
@@ -1,3 +1,4 @@
+from typing import Any
 from dataclasses import dataclass, field
 from datetime import datetime
 
@@ -5,3 +6,4 @@ from datetime import datetime
 @dataclass(frozen=True)
 class CoreMessageDefault:
     timestamp: datetime = field(default_factory=datetime.now)
+    chat: Any = None
