+++ telegram_framework/messages/reply/reply_base.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
-from telegram_framework.messages.chat_message_base import ChatMessageBase
+from telegram_framework.messages.text_message import MessageBase
 
 
 @dataclass(frozen=True)
-class ReplyBase(ChatMessageBase):
-    message: ChatMessageBase
+class ReplyBase(MessageBase):
+    message: MessageBase
